% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_one.R
\name{as.character.zero_one}
\alias{as.character.zero_one}
\title{not clear how to use this. I want an easy way to label the vertical axis of graphs
where the y-axis is a zero-one variable.}
\usage{
\method{as.character}{zero_one}(x, ...)
}
\description{
not clear how to use this. I want an easy way to label the vertical axis of graphs
where the y-axis is a zero-one variable.
}
\examples{
P <- gf_jitter(zero_one(outcome) ~ age, color=~smoker, data = Whickham) \%>\% label_zero_one() \%>\% gf_labs(y="Smoker")
mod <- mod_train(zero_one(outcome) ~ age + smoker, data = Whickham)
fun <- makeFun(mod)
P
P \%>\% mosaicCalc::slice_plot(fun(age, smoker="Yes") ~ age)
}
