% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{coefficients}
\alias{coefficients}
\title{Return the coefficients from a model in the form of a data frame}
\usage{
coefficients(mod)
}
\description{
This is a wrapper around stats::coefficients()
configured to put the name of each coefficient
in the "terms" column.
}
