% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_model_residuals.R
\name{compare_model_residuals}
\alias{compare_model_residuals}
\title{Compare models based on the size of their residuals}
\usage{
compare_model_residuals(
  source,
  ...,
  n = 500,
  measure = c("RMS", "SS", "R2"),
  testing = c("in-sample", "out-of-sample", "R2")
)
}
\arguments{
\item{source}{A data frame or a DAG (as made by dag_make())}

\item{n}{Sample size from DAG.}

\item{measure}{Either of \code{"RMS"} (the default) or \code{"SS"} (sum of squares)}

\item{\ldots}{One or more model formulas, all with the same response variable.}

\item{in_sample}{Boolean flag: whether to use in-sample testing. Default: FALSE.}
}
\description{
Compare a set of models, all with the same response variable, using either
RMS residuals or the sum-of-square residuals
}
\examples{
compare_model_residuals(mtcars, mpg ~ 1, mpg ~ hp, mpg ~ hp + wt, measure="SS", in_sample=TRUE)
compare_model_residuals(dag07, c ~ 1, c ~ a, c ~ a + b, c ~ a + b + d, measure="SS", in_sample=FALSE)
}
