% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_rms_error}
\alias{compare_rms_error}
\title{Compare models}
\usage{
compare_rms_error(DAG, ..., n = 500, in_sample = FALSE)
}
\arguments{
\item{DAG}{A DAG, as made by dag_make()}

\item{n}{Sample size from DAG.}

\item{in_sample}{Boolean flag: whether to use in-sample testing. Default: FALSE.}

\item{\ldots}{One or more model formulas, all with the same response variable.}
}
\description{
Compare a set of models, all with the same response variable, for RMS prediction error.
}
