% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_one.R
\name{label_zero_one}
\alias{label_zero_one}
\title{need to write documentation for the graphics command \code{label_zero_one()}}
\usage{
label_zero_one(P)
}
\description{
need to write documentation for the graphics command \code{label_zero_one()}
}
