% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_eval.R
\name{model_eval}
\alias{model_eval}
\title{Evaluate a model on an input data set.}
\usage{
model_eval(
  mod,
  input = NULL,
  data = NULL,
  ...,
  type = c("response", "link"),
  interval = c("prediction", "confidence", "none"),
  level = 0.95,
  skeleton = FALSE
)
}
\description{
Evaluate a model on an input data set.
}
