% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plot.R
\name{model_plot}
\alias{model_plot}
\title{Plot model output for representative levels of explanatory variables}
\usage{
model_plot(mod, data = NULL)
}
\description{
Plot model output for representative levels of explanatory variables
}
