% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_skeleton.R
\name{model_skeleton}
\alias{model_skeleton}
\title{Convert a data frame to a skeleton}
\usage{
model_skeleton(mod, data = NULL, ncont = 10)
}
\arguments{
\item{mod}{A fitted model \strong{or} a tilde expression describing a model structure, e.g. \code{outcome ~ vara+varb}.}

\item{data}{a data frame}

\item{ncont}{number of levels at which to represent a continuous variable that comes first in the formula}

\item{tilde}{the model specification as a tilde expression}

\item{spreadn}{integer, number of the explanatory variables to give multiple levels}

\item{\ldots}{specific levels to use for variables}

\item{nlevels}{number of levels at which to represent other variables}
}
\description{
Only explanatory variables are included in the result. Use \code{expand.grid()} on
the output to create a data frame with the skeleton values.
}
