% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_summary.R
\name{regression_summary}
\alias{regression_summary}
\title{Generate a regression report in the frm of a data frame.}
\usage{
regression_summary(model, ...)
}
\arguments{
\item{model}{A model as produced by \code{lm()}, \code{glm()}, and so on}
}
\description{
Generate a regression report in the frm of a data frame.
}
\details{
This is simply a wrapper around \code{broom::tidy()} used to
emphasize to students that the results are a summary in the form of a regression
report, similar to the summaries produced by \code{confint()}, \code{coefficients()}, etc.
}
