% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_skeleton.R
\name{data_skeleton}
\alias{data_skeleton}
\title{Convert a data frame to a skeleton}
\usage{
data_skeleton(data, tilde, ..., ncont = 100, nlevels = 3)
}
\arguments{
\item{data}{a data frame}

\item{tilde}{the model specification as a tilde expression}

\item{ncont}{number of levels at which to represent a continuous variable that comes first in the formula}

\item{nlevels}{number of levels at which to represent other variables}

\item{\ldots}{specific levels to use for variables}
}
\description{
Only explanatory variables are included in the result.
}
